flask_proj/requirements.txt
flask
pytest
SnowNLP
bs4
line-bot-sdk